requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Deprecated==1.2.10
fastapi==0.63.0
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.1
pydantic==1.7.3
PyGithub==1.54.1
PyJWT==1.7.1
python-dotenv==0.15.0
regex==2020.11.13
requests==2.25.1
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
wrapt==1.12.1
